Split eval module into main and instructions
Also get rid of the Eval trait. It was no longer buying us much.
