+++ flatgfa-sh/src/eval/mod.rs
@@ -1,9 +1,10 @@
-use crate::ir::{self, Op, Resource, ResourceKind};
-use bstr::BStr;
+mod instr;
+
+use crate::ir::{self, Resource, ResourceKind};
 use enum_map::EnumMap;
 use flatgfa::FlatGFA;
 use flatgfa::flatbed::HeapBEDStore;
-use flatgfa::{self, emit::Emit, flatgfa::HeapGFAStore, memfile, ops};
+use flatgfa::{self, emit::Emit, flatgfa::HeapGFAStore, memfile};
 use memmap::Mmap;
 use std::ffi::OsStr;
 use std::fs::File;
@@ -204,152 +205,9 @@ impl Output {
     }
 }
 
-trait Eval {
-    fn eval(&self, env: &mut Env);
-}
-
-impl Eval for ir::Instr {
-    fn eval(&self, env: &mut Env) {
-        match &self.op {
-            Op::NodeDepth => node_depth(env, self.input, self.output),
-            Op::PathDepth { path } => path_depth(env, self.input, self.output, path),
-            Op::Exec { command, args } => exec(env, self.input, self.output, command, args),
-            Op::ParseGFA => parse_gfa(env, self.input, self.output),
-            Op::MapFile => map_file(env, self.input, self.output),
-            Op::ParseBED => parse_bed(env, self.input, self.output),
-            Op::MakeWindows { size } => make_windows(env, self.input, self.output, *size),
-            Op::OdgiView => odgi_view(env, self.input, self.output),
-        }
-    }
-}
-
-fn node_depth(env: &mut Env, input: Resource, output: Resource) {
-    let mut out = env.output(output);
-    let gfa = env.flatgfa(input);
-    let (depths, uniq_depths) = ops::depth::seg_depth(&gfa);
-    out.emit(ops::depth::SegDepth {
-        gfa: &gfa,
-        depths,
-        uniq_depths,
-    })
-    .unwrap();
-}
-
-fn path_depth(env: &mut Env, input: Resource, output: Resource, path: &Option<String>) {
-    let mut out = env.output(output);
-    let gfa = env.flatgfa(input);
-    if let Some(path_name) = &path {
-        // TODO More elegantly handle missing paths.
-        let path_id = gfa
-            .find_path(path_name.as_bytes().into())
-            .expect("no such path found");
-        let (lengths, depths) = ops::depth::path_depth(&gfa, std::iter::once(path_id));
-        out.emit(ops::depth::PathDepth {
-            gfa: &gfa,
-            depths,
-            lengths,
-            paths: std::iter::once(path_id),
-        })
-        .unwrap();
-    } else {
-        let (lengths, depths) = ops::depth::path_depth(&gfa, gfa.paths.ids());
-        out.emit(ops::depth::PathDepth {
-            gfa: &gfa,
-            depths,
-            lengths,
-            paths: gfa.paths.ids(),
-        })
-        .unwrap();
-    }
-}
-
-fn exec(env: &mut Env, input: Resource, output: Resource, command: &String, args: &[String]) {
-    env.run_cmd(&command, &args, input, output);
-}
-
-fn parse_gfa(env: &mut Env, input: Resource, output: Resource) {
-    use flatgfa::parse::Parser;
-
-    let store = match env.input(input) {
-        Input::File(file) => Parser::for_heap().parse_mem(file.as_ref()),
-        Input::Stdin(stream) => Parser::for_heap().parse_stream(stream),
-        Input::Pipe(stream) => Parser::for_heap().parse_stream(stream),
-    };
-
-    env.gfa_stores[output] = Some(store);
-}
-
-fn map_file(env: &mut Env, input: Resource, output: Resource) {
-    if let ResourceKind::File = input.kind {
-        let mmap = memfile::map_file(env.file_name(input));
-        env.mmaps[output] = Some(mmap);
-    } else {
-        panic!("can only map actual files");
-    }
-}
-
-fn parse_bed(env: &mut Env, input: Resource, output: Resource) {
-    use flatgfa::flatbed::BEDParser;
-
-    let store = match env.input(input) {
-        Input::File(file) => BEDParser::for_heap().parse_mem(file.as_ref()),
-        Input::Stdin(stream) => BEDParser::for_heap().parse_stream(stream),
-        Input::Pipe(stream) => BEDParser::for_heap().parse_stream(stream),
-    };
-
-    env.bed_stores[output] = Some(store);
-}
-
-struct WindowsBed<'a> {
-    name: &'a BStr,
-    start: u64,
-    end: u64,
-    size: u64,
-}
-
-impl<'a> Emit for WindowsBed<'a> {
-    fn emit(self, f: &mut impl std::io::Write) -> io::Result<()> {
-        let mut pos = self.start;
-        while pos < self.end {
-            let end = (pos + self.size).min(self.end);
-            writeln!(f, "{}\t{}\t{}", self.name, pos, end)?;
-            pos = end;
-        }
-        Ok(())
-    }
-}
-
-fn make_windows(env: &mut Env, input: Resource, output: Resource, size: usize) {
-    let store = env.bed_stores[input].take().unwrap();
-    let bed = store.as_ref();
-    let mut out = env.output(output);
-    for entry in bed.entries.all() {
-        out.emit(WindowsBed {
-            name: bed.get_name_of_entry(entry),
-            start: entry.start,
-            end: entry.end,
-            size: size.try_into().unwrap(),
-        })
-        .unwrap();
-    }
-}
-
-fn odgi_view(env: &mut Env, input: Resource, output: Resource) {
-    let og_file = env.file_name(input).to_string();
-    env.run_cmd(
-        "odgi",
-        &["view", "-g", "-i", &og_file],
-        Resource {
-            kind: ResourceKind::Stdin,
-            index: 0,
-        },
-        output,
-    )
-}
-
 pub fn run(prog: ir::Program) {
     let mut env = Env::new(prog.file_names, prog.rsrc_counts);
-    for op in prog.instrs {
-        op.eval(&mut env);
+    for i in prog.instrs {
+        instr::eval(&mut env, &i);
     }
 }
